logistics-data-pipeline/requirements.txt
# Airflow
apache-airflow==2.10.3

# Data Processing
pandas==2.1.4
numpy==1.26.3

# API & HTTP
requests==2.31.0
urllib3==2.1.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
# sqlalchemy version managed by airflow

# Testing
pytest==7.4.4
pytest-mock==3.12.0
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
